// src/components/ui-kit/checkbox/types.ts
export enum CheckboxProportion {
  Small = "small",
  Medium = "medium",
  Large = "large",
}

export enum CheckboxVariant {
  Success = "success",
  Warning = "warning",
  Error = "error",
  Primary = "primary",
  Secondary = "secondary",
}

export enum CheckboxShape {
  Rounded = "rounded",
  Circle = "circle",
}

export enum CheckboxBorder {
  Light = "light",
  Medium = "medium",
  Bold = "bold",
  None = "none",
}